backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "argon2-cffi>=25.1.0",
    "dj-database-url>=3.0.1",
    "django>=5.2.7",
    "django-cors-headers>=4.9.0",
    "django-filter>=25.2",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "pytest-django>=4.10.0",
    "psycopg2-binary>=2.9.11",
    "psycopg[binary]>=3.2.10",
    "pytest>=8.4.2",
    "python-dotenv>=1.1.1",
    "django-extensions>=4.1",
    "httpx>=0.28.1",
    "pytest-xdist>=3.8.0",
    "pytest-rerunfailures>=16.1",
    "pytest-randomly>=4.0.1",
    "gunicorn>=23.0.0",
    "whitenoise>=6.11.0",
    "django-import-export>=4.3.10",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
